requirements.txt
pyyaml
requests
pandas
scipy
openai
xmltodict
matplotlib
plotly
# sklearn